pycharmworkouts-master/requirements.txt
mysql-connector-python
python-dotenv
pandas
openpyxl
fastapi
uvicorn
requests